fix: arrow closure flag leak + method.rs over 400 lines
1. Arrow functions now save/restore use_state_for_this flag
   Prevents `state` from leaking into nested closures in handlers

2. Moved map_status_code + build_doc_comment to routing.rs
   method.rs: 426 → 388 lines (under 400 limit)

3. Tightened tier4 test assertion (removed dead fallback branch)

179 tests passing. Zero regressions.

Closes #84
